store a string for self.env.host to work around the problem when build
when configure finishes, it stores the env values. these values are
restored by 'eval()' when build begins. The problem with the method is
that it eval()'s the expression in different contexts, which has
different globals() and locals() from the ones when they were stored. To
work around this problem, only basic types are stored in 'env', such as
str, list, dict etc. which can be eval()'ed without regarding to the
globals() and locals()